requirements.txt
# Core LangChain
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langchain-core==0.3.0

# Vector Database with full dependencies
chromadb==0.4.22
onnxruntime==1.16.0
sentence-transformers==2.2.2

# Additional ChromaDB dependencies
chroma-hnswlib==0.7.3
tokenizers==0.15.0

# Web Framework
streamlit==1.29.0

# Google AI
google-generativeai==0.8.0

# Document Processing
PyPDF2==3.0.1
PyMuPDF==1.23.26
pdfplumber==0.10.3
pycryptodome==3.19.0
python-docx==1.1.0
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Scraping
requests==2.31.0
selenium==4.16.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Image Processing
Pillow==10.1.0

# Async support
aiofiles==23.2.1

# Logging
loguru==0.7.2